Implement own java ast to replace treesitter

+++ zip_util/src/lib.rs
@@ -1,3 +1,6 @@
+#![deny(warnings)]
+#![deny(clippy::unwrap_used)]
+#![deny(clippy::redundant_clone)]
 use rc_zip_tokio::{ReadZip, rc_zip::parse::EntryKind};
 use std::{
     fs::File,
@@ -51,10 +54,10 @@ pub async fn extract_jar(jar: &PathBuf, source_dir: &Path) -> Result<(), ZipUtil
                             .expect("all full entry paths should have parent paths"),
                     )
                     .map_err(ZipUtilError::IO)?;
-                    if let Ok(metadata) = std::fs::symlink_metadata(&path) {
-                        if metadata.is_file() {
-                            std::fs::remove_file(&path).map_err(ZipUtilError::IO)?;
-                        }
+                    if let Ok(metadata) = std::fs::symlink_metadata(&path)
+                        && metadata.is_file()
+                    {
+                        std::fs::remove_file(&path).map_err(ZipUtilError::IO)?;
                     }
 
                     let mut src = String::new();
